Finalizacion CRUD categorias

+++ frontend/src/app/product/categories/categories.component.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
-import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { CategoriesService } from '../../core/services/categories.service';
 import { Category } from '../../core/models/categories.interface';
 
@@ -10,34 +10,44 @@ import { Category } from '../../core/models/categories.interface';
 @Component({
   selector: 'app-categories',
   standalone: true,
-  imports: [CommonModule, ReactiveFormsModule],
+  imports: [CommonModule, ReactiveFormsModule, FormsModule],
   templateUrl: './categories.component.html',
   styleUrl: './categories.component.css'
 })
 export class CategoriesComponent implements OnInit {
 categoryForm: FormGroup=new FormGroup({});
 categories: Category[] = [];
+selectedCategory: Category | null = null;
 
 
 
 constructor(
   private fb: FormBuilder, 
   private categoriesService: CategoriesService,
-){}
+
+  ) {
+    this.categoryForm = this.fb.group({
+      nombreCategoria: [''],
+      descripcionCategoria: ['']
+    });
+   }
+
 
 ngOnInit(): void {
-this.categoryForm = this.fb.group({
-nombreCategoria: ['', Validators.required],
-descripcionCategoria: ['', Validators.required]
-})
-this.loadCategories();
+  this.loadCategories();
 }
 
-openModal(){
-  const modalDiv = document.getElementById('editCategoria');
+openModal(modalId: string, category: Category): void{
+  this.selectedCategory = category;
+  if (category) {
+    this.categoryForm.patchValue({
+      nombreCategoria: category.nombreCategoria,
+      descripcionCategoria: category.descripcionCategoria
+    });
+  }
+  const modalDiv = document.getElementById(modalId);
   if(modalDiv != null){
     modalDiv.style.display='block';
-    
   }
 }
 
@@ -50,17 +60,14 @@ closeModal(modalId: string){
   if (backdrop != null) {
     backdrop.parentNode?.removeChild(backdrop);
   }
+  this.selectedCategory= null;
 }
-
-
 
 loadCategories(): void {
   this.categoriesService.getAllCategories().subscribe((categories : Category[]) => {
     this.categories = categories;
   });
 }
-
-
 
 addCategory() {
   if (this.categoryForm.valid) {
@@ -78,14 +85,30 @@ addCategory() {
 }
 
 
-editCategory(category: Category) {
+editCategory(): void {
+  if (this.selectedCategory) {
+    const updatedCategory: Category = {
+      ...this.selectedCategory,
+      ...this.categoryForm.value
+    };
+    this.categoriesService.editCategory(updatedCategory).subscribe(() => {
+      alert('Categoría actualizada');
+      this.closeModal('editCategoria');
+      this.loadCategories();
+      this.categoryForm.reset();
+    });
+  }
 }
 
-deleteCategory(category: Category) {
-  this.categoriesService.deleteCategory(category).subscribe(() => {
-    alert('Categoria eliminada');
-    this.ngOnInit();
-  });
+deleteCategory(category: Category | null, modalId: string) {
+  if(category){
+      this.categoriesService.deleteCategory(category).subscribe(() => {
+        alert('Categoria eliminada');
+        this.ngOnInit();
+        this.closeModal(modalId);
+        this.categoryForm.reset();
+      });
+  }
 }
 
 }
